fix(laptop): updating ssh to avoid warning and limiting karabiner popups following build

+++ home/modules/karabiner/default.nix
@@ -13,8 +13,10 @@ in {
   config = lib.mkIf cfg.enable {
     home.activation.karabiner = lib.hm.dag.entryAfter ["writeBoundary"] ''
       run mkdir -p "${config.xdg.configHome}/karabiner"
-      run cp -f "${karabinerConfig}" "${config.xdg.configHome}/karabiner/karabiner.json"
-      run chmod u+rw "${config.xdg.configHome}/karabiner/karabiner.json"
+      if ! cmp -s "${karabinerConfig}" "${config.xdg.configHome}/karabiner/karabiner.json"; then
+        run cp -f "${karabinerConfig}" "${config.xdg.configHome}/karabiner/karabiner.json"
+        run chmod u+rw "${config.xdg.configHome}/karabiner/karabiner.json"
+      fi
     '';
   };
 }
